Add show button for example 2.13 Date for countries
To add show button feature, added FilteredResult component to manage result page.
It receives filteredCountries as props, as well as clickedCountry state.
Based on length of filteredCountries, it returns differnet result.
Also when show button is clicked, clickedCountry state is updated with clicked country data.
That state is given to CountryData component so that it renders that specific country information.